+++ src/main/java/nordmods/uselessreptile/common/entity/base/HeadMountDragon.java
@@ -1,4 +1,7 @@
 package nordmods.uselessreptile.common.entity.base;
 
 public interface HeadMountDragon {
+    default URDragonEntity asURDragon() {
+        return (URDragonEntity) this;
+    }
 }