+++ src/main/java/nordmods/uselessreptile/mixin/common/AbstractBlockMixin.java
@@ -9,7 +9,8 @@ import net.minecraft.util.Hand;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import nordmods.uselessreptile.common.entity.RiverPikehornEntity;
+import nordmods.uselessreptile.common.entity.base.HeadMountDragon;
+import nordmods.uselessreptile.common.entity.base.URDragonEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -19,7 +20,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class AbstractBlockMixin {
     @Inject(method = "onUse", at = @At("HEAD"), cancellable = true)
     private void putDragonAssOff(BlockState state, World world, BlockPos pos, PlayerEntity player, BlockHitResult hit, CallbackInfoReturnable<ActionResult> cir) {
-        if (player.isSneaking() && player.getFirstPassenger() instanceof RiverPikehornEntity dragon ) {
+        if (player.isSneaking() && player.getFirstPassenger() instanceof HeadMountDragon headMountDragon && headMountDragon instanceof URDragonEntity dragon) {
             dragon.stopRiding();
             dragon.setPosition(pos.up().toCenterPos());
             cir.setReturnValue(ActionResult.SUCCESS);
@@ -28,7 +29,7 @@ public abstract class AbstractBlockMixin {
 
     @Inject(method = "onUseWithItem", at = @At("HEAD"), cancellable = true)
     private void putDragonAssOff(ItemStack stack, BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit, CallbackInfoReturnable<ActionResult> cir){
-        if (player.isSneaking() && player.getFirstPassenger() instanceof RiverPikehornEntity dragon ) {
+        if (player.isSneaking() && player.getFirstPassenger() instanceof HeadMountDragon headMountDragon && headMountDragon instanceof URDragonEntity dragon) {
             dragon.stopRiding();
             dragon.setPosition(pos.up().toCenterPos());
             cir.setReturnValue(ActionResult.SUCCESS);