+++ src/main/java/nordmods/uselessreptile/mixin/common/PlayerEntityMixin.java
@@ -1,11 +1,16 @@
 package nordmods.uselessreptile.mixin.common;
 
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.nbt.NbtCompound;
+import net.minecraft.nbt.NbtElement;
+import net.minecraft.server.world.ServerWorld;
 import net.minecraft.world.World;
-import nordmods.uselessreptile.common.util.LightningChaserSpawnTimer;
+import nordmods.uselessreptile.common.util.duck.HeadMountDragonOwner;
+import nordmods.uselessreptile.common.util.duck.LightningChaserSpawnTimer;
+import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
@@ -13,8 +18,9 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(PlayerEntity.class)
-public abstract class PlayerEntityMixin extends LivingEntity implements LightningChaserSpawnTimer {
+public abstract class PlayerEntityMixin extends LivingEntity implements LightningChaserSpawnTimer, HeadMountDragonOwner {
     @Unique private int lightningChaserSpawnCooldown = 0;
+    @Unique private NbtCompound headMountDragon = new NbtCompound();
     protected PlayerEntityMixin(EntityType<? extends LivingEntity> entityType, World world) {
         super(entityType, world);
     }
@@ -22,11 +28,15 @@ public abstract class PlayerEntityMixin extends LivingEntity implements Lightnin
     @Inject(method = "writeCustomDataToNbt(Lnet/minecraft/nbt/NbtCompound;)V", at = @At("TAIL"))
     private void writeToNbt(NbtCompound nbt, CallbackInfo ci) {
         nbt.putInt("LightningChaserSpawnCooldown", useless_reptile$getTimer());
+        if (!headMountDragon.isEmpty()) nbt.put("HeadMountDragon", headMountDragon);
     }
 
     @Inject(method = "readCustomDataFromNbt(Lnet/minecraft/nbt/NbtCompound;)V", at = @At("TAIL"))
     private void readFromNbt(NbtCompound nbt, CallbackInfo ci) {
         useless_reptile$setTimer(nbt.getInt("LightningChaserSpawnCooldown"));
+        if (nbt.contains("HeadMountDragon", NbtElement.COMPOUND_TYPE)) {
+            setHeadMountDragon(nbt.getCompound("HeadMountDragon"));
+        }
     }
 
     public int useless_reptile$getTimer() {
@@ -39,5 +49,26 @@ public abstract class PlayerEntityMixin extends LivingEntity implements Lightnin
     @Inject(method = "tick", at = @At("TAIL"))
     private void tickTimer(CallbackInfo ci) {
         if (useless_reptile$getTimer() > 0) useless_reptile$setTimer(useless_reptile$getTimer() - 1);
+    }
+
+    @Override
+    public void setHeadMountDragon(@NotNull NbtCompound state) {
+        headMountDragon = state;
+    }
+
+    @Override
+    @NotNull
+    public NbtCompound getHeadMountDragon() {
+        return headMountDragon;
+    }
+
+    @Inject(method = "remove", at = @At("TAIL"))
+    private void removeHeadMountDragon(RemovalReason reason, CallbackInfo ci) {
+        if (!headMountDragon.isEmpty() && getWorld() instanceof ServerWorld world) {
+            if (!reason.shouldDestroy()) {
+                Entity dragon = world.getEntity(headMountDragon.getUuid("UUID"));
+                if (dragon != null) dragon.remove(reason);
+            }
+        }
     }
 }